Remove failing getProjectByHandler call
